requirements.txt
python-telegram-bot
deep_translator